♻️ refactor(session): 基盤をイベント型フレームワーク化＋voting を外部型に切り出し
「基盤にイベント型を焼き込む」のをやめ、基盤＝汎用フレームワーク、イベント型＝外部の
サーバー＋クライアント・ペアという構成に。新しい型（チュートリアル/カードゲーム）は
ServerSessionType＋ClientSessionType のペアを1パッケージ足すだけで追加できる。

- session-protocol: 封筒を型非依存に汎用化（SessionType=string、public/data/action/config
  は unknown）。サーバー拡張契約 ServerSessionType（自己記述する type id 付き）を移設。
- usketch-plugin-session: フレームワーク化。汎用 session-client（act 中心）＋ HUD パネル外枠
  ＋クライアント型レジストリ。createSessionPlugin({ types }) が各 ClientSessionType へ
  renderCard/renderCreateForm を委譲。SessionManager は型レジストリを注入で受け取る。
  基盤はもう voting を知らない。
- 新パッケージ usketch-session-voting: 投票を最初の外部イベント型として実装。./server
  （votingServerType）と ./client（votingClientType）の2エントリ。apps/server と apps/web が
  それぞれ import。サーバーバンドルに React は入らない。
- apps/server: SessionManager に types:[votingServerType] を注入、旧 session-types/voting.ts 削除。
- apps/web: createSessionPlugin に types:[votingClientType] を渡す。

挙動・単体テスト（サーバー13・クライアント10）は不変。ブラウザ実機で汎用シェル経由の
作成→投票→締切→終了を確認。
