+++ apps/server/package.json
@@ -21,6 +21,7 @@
 		"@edv4h/usketch-plugin-server-comments": "workspace:*",
 		"@edv4h/usketch-server-core": "workspace:*",
 		"@edv4h/usketch-session-protocol": "workspace:*",
+		"@edv4h/usketch-session-voting": "workspace:*",
 		"@edv4h/usketch-shared": "workspace:*",
 		"@edv4h/usketch-sync": "workspace:*",
 		"@hono/zod-validator": "0.9.0",
