+++ plugins/usketch-plugin-session/src/index.ts
@@ -2,3 +2,8 @@ export type { SessionPluginOptions } from "./plugin.js";
 export { createSessionPlugin } from "./plugin.js";
 export type { SessionClient, SessionClientState } from "./session-client.js";
 export { createSessionClient } from "./session-client.js";
+export type {
+	ClientSessionType,
+	SessionCardContext,
+	SessionCreateContext,
+} from "./session-type.js";
